:memo: Add renaming transforms
This adds the python_minifier.rename package for safely changing the names
used in a python module.

This is used by the new rename_locals and rename_globals options of the minify
and pyminify command to shorten names.

The hoist_literals transform has been rewritten to use the new rename package
for newly introduced names.