Update github actions versions
GitHub decided to use nodejs for their github actions, such as checkout/upload-artifact/download-artifact etc.
This is problematic when running actions jobs inside containers, as node needs to get in there somehow.
The actions runner rams a nodejs runtime into your containers, whether it works there or not.

GitHub are threatening to use a newer nodejs runtime, which certainly does NOT work in some of the older python containers.

To address this there is a new composite action called 'run-in-container', which tries to approximate the same behaviour
as running jobs inside containers, but for a single step. This means that the GitHub provided actions in the job can now
run outside the test container.